Bug fixes and comments
maxVal name changed because f typo (from maxVall)
Got rid of some useless lines
Fixed small bug from calculating action. Wrong value was incremented in loop.
Optimized part of code where the desired action was found.
Fixed bug from some variables that were int type instead of float.
Commented parts of the action calculation function.

+++ src/creature.h
@@ -14,7 +14,7 @@ typedef struct Creature {
     int actionSinkCount[ACTION_NEURONS];
     Genome* brainsInnerNeuronsSink[INNER_NEURONS][BRAIN_SIZE];
     Genome* brainsActionNeuronsSink[ACTION_NEURONS][BRAIN_SIZE];
-    unsigned int innerBufferedValues[INNER_NEURONS];
+    float innerBufferedValues[INNER_NEURONS];
     
 } Creature;
